requirements.txt
tensorflow==2.0.0
tensorflow-hub==0.10.0
bert-for-tf2==0.14.7
sentencepiece==0.1.94
matplotlib==3.3.4
seaborn==0.11.2
pydot==1.4.2
graphviz==0.19.1
pandas==1.1.5
scikit-learn==0.24.2
nltk==3.4.5
gensim==3.8.1
